修改了一些解码器长度溢出的问题（还有其他问题）。 添加谷歌Guava来管理缓存，缓存部分还很简单。 测试了请求分发的部分，修改了一些关于注入的bug。

+++ gameserver/src/main/java/com/wan37/gameServer/server/handler/MessageEncoder.java
@@ -1,6 +1,6 @@
 package com.wan37.gameServer.server.handler;
 
-import com.wan37.gameServer.common.Message;
+import com.wan37.common.entity.Message;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToByteEncoder;
@@ -21,8 +21,8 @@ public class MessageEncoder extends MessageToByteEncoder<Message> {
         out.writeByte(-128)
                 .writeInt(length)
                 .writeInt(msg.getMsgId())
-                .writeInt(msg.getType())
-                .writeInt(msg.getFlag())
+                .writeByte(msg.getType())
+                .writeByte(msg.getFlag())
                 .writeBytes(msg.getContent());
 
         ctx.flush();
